+++ superset/viz_sip38.py
@@ -476,7 +476,7 @@ class BaseViz:
         stacktrace = None
         df = None
         cached_dttm = datetime.utcnow().isoformat().split(".")[0]
-        if cache_key and cache_manager.data_cache and not self.force:
+        if cache_key and not self.force:
             cache_value = cache_manager.data_cache.get(cache_key)
             if cache_value:
                 stats_logger.incr("loading_from_cache")
